autogenerate pages for events imported by manifest
